Sort of works
The problem is that in this range of data (mean(|x|) approx. 0.3), polynomials such as x^3 and x^4 are so close to 0 that their contribution barely even matters. If the data does not have any higher orders than x^2, then the estimation procedure works fine. This might necessitate a clever form of regularisation.