Results: Elbow Curves
I clustered the sector-data with a regular kmeans approach because the neuralisation for the bipartite-split kmeans seems to work differently (although qualitatively, the resulting plots in Time vs. Cluster look similar). The cluster labels can now be analysed with the XAI-method and the results are depicted at the bottom as pots of the ranked mean XAI-importance of the features. There are some elbow curves visible in the plots, both at the left and right extremes. On the left-hand side, importances below 0 correspond to features that "speak against" the assignment, while on the right-hand side, high importances support the assignment.
The whole code still needs to be cleaned up.