Evaluated the Elbow Curves
Wrote a function to evaluate when the elbow starts in an "elbow curve" by comparing the values to an interpolated linear regression for the first/central data values. This allows us to separate the features with high XAI-values from the rest. The most hyperparameters are the interval of values used for the regression and the number of standard deviations that are necessary to be classified as out-of-prediction.