Change Point Analysis
This script takes the 1D time series of the mean correlations of the returns (in a rolling 42 day time window) and tries to find change points in the mean correlation. It uses the antiCPy package and observes the time series in a rolling window of 1000 measurements, where it detects CPs individually. The points of Maximum Likelihood for the CP position and the overall likelihood are summed over all rolling windows to aggregate it for the final plots. However, the first and last "safetymargin" number of points are disregarded (here 30% each of the window) because the algorithm can behave unstable at these points. Also, all windows with nans or infs are ignored to prevent empty invervals in the aggregated plot.