Change Point Analysis: Shift 1
Now the window is shifted by one observation for each new window. The composite npy results are attached, too. Only 43 of the 4291 windows have to be disregarded because of NaNs/Infs which means that 99% of the data is preserved. Hence, I do not believe that any relevant information is lost by this process.